refactor: Give explicit name to individual action/reducer/saga files
description: Currently every screen action is names actions.js which make it difficult to find particular file. Hence change for home, homeActions.js from actions.js similarly for all screens and same for reducer and saga file
