+++ src/screens/PaymentScreen/PaymentScreen.js
@@ -27,6 +27,7 @@ import { priceSignByCode } from '../../utils/price';
 const PaymentScreen = ({
   payment,
   orderStatus,
+  orderId,
   billingAddress,
   navigation,
   currencyCode,
@@ -44,6 +45,16 @@ const PaymentScreen = ({
     // componentDidUnmount: Reset Payment related logic in Redux
     _resetPaymentState();
   }), []);
+
+  useEffect(() => {
+    if (orderStatus === Status.SUCCESS) {
+      _createQuoteId();
+      navigation.reset({
+        index: 1,
+        routes: [{ name: NAVIGATION_TO_HOME_SCREEN }, { name: NAVIGATION_TO_ORDER_CONFIRMATION_SCREEN, params: { status: Status.SUCCESS, orderId } }],
+      });
+    }
+  }, [orderStatus]);
 
   const placeOrder = () => {
     if (!paymentCode) return;
@@ -149,14 +160,6 @@ const PaymentScreen = ({
     />
   );
 
-  if (orderStatus === Status.SUCCESS) {
-    _createQuoteId();
-    navigation.reset({
-      index: 1,
-      routes: [{ name: NAVIGATION_TO_HOME_SCREEN }, { name: NAVIGATION_TO_ORDER_CONFIRMATION_SCREEN, params: { status: Status.SUCCESS } }],
-    });
-  }
-
   return (
     <GenericTemplate
       scrollable={false}
@@ -184,6 +187,7 @@ PaymentScreen.propTypes = {
   billingAddress: PropTypes.object,
   payment: PropTypes.object,
   orderStatus: PropTypes.oneOf(Object.values(Status)).isRequired,
+  orderId: PropTypes.string,
   placeCartOrder: PropTypes.func.isRequired,
   createQuoteId: PropTypes.func.isRequired,
   currencyCode: PropTypes.string.isRequired,
@@ -196,10 +200,11 @@ PaymentScreen.propTypes = {
 PaymentScreen.defaultProps = {
   billingAddress: {},
   payment: [],
+  orderId: -1,
 };
 
 const mapStateToProps = ({ checkout, cart, magento }) => {
-  const { payment, orderStatus } = checkout;
+  const { payment, orderStatus, orderId } = checkout;
   const { cart: { billing_address: billingAddress } } = cart;
   const {
     base_currency_symbol: baseCurrencySymbol,
@@ -210,6 +215,7 @@ const mapStateToProps = ({ checkout, cart, magento }) => {
   return {
     payment,
     orderStatus,
+    orderId,
     billingAddress,
     baseCurrencySymbol,
     currencyCode,
