+++ src/store/actions.js
@@ -1,10 +1,10 @@
-export * from './magento/actions';
-export * from './home/actions';
-export * from './categoryTree/actions';
-export * from './auth/actions';
-export * from './account/actions';
-export * from './cart/actions';
-export * from './categoryList/actions';
-export * from './checkout/actions';
-export * from './product/actions';
-export * from './search/actions';
+export * from './account/accountActions';
+export * from './auth/authActions';
+export * from './cart/cartActions';
+export * from './magento/magentoActions';
+export * from './home/homeActions';
+export * from './categoryTree/categoryTreeActions';
+export * from './categoryList/categoryListActions';
+export * from './checkout/checkoutActions';
+export * from './product/productActions';
+export * from './search/searchActions';
